Handle errors connecting to remote switch
Connections to the remote switch would occasionally fail with an OSError:

    File "urequests.py", line 60, in request
    OSError: [Errno 103] ECONNABORTED

This commit transforms the turn_on/turn_off methods into coroutines, and
wraps the requests.get call in a try/except block inside a while loop. We
call the commands asynchronously so the rest of the code can continue
running while we retry connections to the switch.